feat: 카메라 페이지에서 포워딩 페이지로 전환 후 전송 처리
