+++ src/kiosk/pages/Forwarding/index.tsx
@@ -1,32 +1,63 @@
-import * as C from '@/allFiles'
-import * as S from './style'
+import * as C from '@/allFiles';
+import * as S from './style';
 
-import { FaPaperPlane } from "react-icons/fa";
+import { FaPaperPlane } from 'react-icons/fa';
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
+import { mainApi } from '@/__api__/axiosInstance';
+import { TextToSpeech } from '@/kiosk/common/services/textToSpeech';
 
 const Forwarding = () => {
-  const [progress, setProgress] = useState(0); 
-  const navigate = useNavigate()
-  // 임의로 움직임 추가
+  const [progress, setProgress] = useState(0);
+  const navigate = useNavigate();
+
   useEffect(() => {
+    const sendImage = async () => {
+      try {
+        const imageData = sessionStorage.getItem('capturedImage'); // 세션에서 꺼냄
+        if (!imageData) throw new Error('전송할 이미지가 없습니다.');
+
+        const base64String = imageData.split(',')[1];
+        const byteCharacters = atob(base64String);
+        const byteNumbers = new Array(byteCharacters.length);
+        for (let i = 0; i < byteCharacters.length; i++) {
+          byteNumbers[i] = byteCharacters.charCodeAt(i);
+        }
+        const byteArray = new Uint8Array(byteNumbers);
+        const blob = new Blob([byteArray], { type: 'image/png' });
+
+        const formData = new FormData();
+        formData.append('conversationId', '1');
+        formData.append('file', blob, 'photo.png');
+
+        await mainApi.post('/photo', formData, {
+          headers: { 'Content-Type': 'multipart/form-data' },
+        });
+
+        TextToSpeech('사진이 성공적으로 전송되었습니다.');
+      } catch (err) {
+        console.error('전송 실패:', err);
+        TextToSpeech('사진 전송에 실패했습니다.');
+        alert('전송 실패');
+      }
+    };
+
+    sendImage();
+
+    // 프로그레스바 애니메이션
     const timer = setInterval(() => {
-      setProgress((prev) => {
+      setProgress(prev => {
         const next = prev + 5;
         if (next >= 100) {
           clearInterval(timer);
-
-          setTimeout(() => {
-            navigate('/forwarding/success'); 
-          }, 500);
-          // 완료 후 처리
+          setTimeout(() => navigate('/forwarding/success'), 500);
         }
         return next;
       });
     }, 100);
 
     return () => clearInterval(timer);
-  }, []);
+  }, [navigate]);
 
   return (
     <S.ForwardingContainer>
@@ -44,5 +75,6 @@ const Forwarding = () => {
       </S.ForwardingWrapper>
     </S.ForwardingContainer>
   );
-}
-export default Forwarding
+};
+
+export default Forwarding;
